Improve performance

+++ project/contracts/helpers/ModMathUser.sol
@@ -4,11 +4,11 @@ pragma solidity 0.8.21;
 import "../ModMath.sol";
 
 contract ModMathUser {
-    function findInverse(uint256 p, uint256 q, uint256 start) external pure returns (Inverse memory) {
+    function findInverse(uint128 p, uint128 q, uint128 start) external pure returns (Inverse memory) {
         return ModMath.findInverse(p, q, start);
     }
 
-    function findSquare(uint256 p, uint256 q, uint256 start) external pure returns (Square memory) {
+    function findSquare(uint128 p, uint128 q, uint128 start) external pure returns (Square memory) {
         return ModMath.findSquare(p, q, start);
     }
 }
